requirements.txt
PyYAML>=6.0,<7.0
beautifulsoup4>=4.12,<5.0
